+++ Changelog.md
@@ -20,6 +20,9 @@
 ####2016-02-20:
  - \#312: Newlines are not accepted in parameter expansions
 
+####2016-02-20:
+ - \#313: More gentle parsing for empty command lists for: for, while, until, if, select.
+
 ####2016-01-06:
  - \#299: ClassNotFound exception in PHPStorm